requirements.txt
PyMuPDF==1.23.14
pdf2image==1.16.3
Pillow==10.1.0
opencv-python==4.8.1.78
layoutparser[ocr]==0.3.4
numpy==1.24.3
pandas==2.1.4
matplotlib==3.8.2
jupyter==1.0.0
ipykernel==6.26.0
pytest==7.4.3
black==23.11.0 